configure some error after merge
